refactor(integration): remove unused components and clean up integration pages
- Deleted obsolete Vue components related to AutoResponder, CodeSnippet, CreateNewForm, ExternalScript, ListAccount, QiscusBannerDoc, TableListChannel, and Widget-related files.
- Improved overall project structure by eliminating unnecessary files, enhancing maintainability and reducing clutter in the integration section.
