--- src/pages/integration/whatsapp/WhatsappBannerDoc.vue
@@ -1,16 +1,0 @@
-<template>
-  <Banner>
-    <p>
-      To learn more about WhatsApp Integration, you can refer to this
-      <a
-        class="text-notification-link font-semibold underline"
-        href="https://documentation.qiscus.com/omnichannel-chat/whatsapp-integration"
-        target="_blank"
-        >Documentation</a
-      >.
-    </p>
-  </Banner>
-</template>
-<script setup lang="ts">
-import Banner from '@/components/common/Banner.vue';
-</script>
